+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v1.1.6-beta.2.240604
 ### 2024/6/4
 - add model detect img category
+- fix bugs on category model detect
 
 ## v1.1.6-beta.1.240531
 ### 2024/5/31